extracted timeline badges ui to `timeline::common::timeline_badge_ui()`
